Delete MCP test by merging into existing test
